增加proto协议支持！

+++ message/message.go
@@ -3,16 +3,16 @@ package message
 var message interface{} = DefaultMessage{}
 
 type Message interface {
-	GetMerge() int
+	GetMerge() int64
 }
 
 // DefaultMessage 必须实现 Message 接口
 type DefaultMessage struct {
-	Merge int
+	Merge int64
 	Body  string
 }
 
-func (d DefaultMessage) GetMerge() int {
+func (d DefaultMessage) GetMerge() int64 {
 	return d.Merge
 }
