+++ router/router.go
@@ -5,31 +5,31 @@ import (
 	"log"
 )
 
-var routerMap = make(map[int]func(msg interface{}) interface{})
+var routerMap = make(map[int64]func(msg interface{}) interface{})
 
 type CmdInfo struct {
-	Cmd      int
-	SubCmd   int
-	CmdMerge int
+	Cmd      int64
+	SubCmd   int64
+	CmdMerge int64
 }
 
 // GetMerge 获取路由
-func GetMerge(cmd, subCmd int) int {
+func GetMerge(cmd, subCmd int64) int64 {
 	return (cmd << 16) + subCmd
 }
 
 // GetCmd 获取主命令
-func GetCmd(merge int) int {
+func GetCmd(merge int64) int64 {
 	return merge >> 16
 }
 
 // GetSubCmd 获取子命令
-func GetSubCmd(merge int) int {
+func GetSubCmd(merge int64) int64 {
 	return merge & 0xFFFF
 }
 
 // AddFunc 添加路由,如果需要返回给客户端数据，需要返回数组字节: []byte
-func AddFunc(merge int, method func(msg interface{}) interface{}) {
+func AddFunc(merge int64, method func(msg interface{}) interface{}) {
 	if routerMap[merge] == nil {
 		routerMap[merge] = method
 	} else {
@@ -38,7 +38,7 @@ func AddFunc(merge int, method func(msg interface{}) interface{}) {
 }
 
 // ExecuteFunc 执行消息
-func ExecuteFunc(merge int, v interface{}) interface{} {
+func ExecuteFunc(merge int64, v interface{}) interface{} {
 	return routerMap[merge](v)
 }
